Top/requirements.txt
h5py
numpy
pandas
Pillow-SIMD
scikit-learn
tensorflow